fix(simulateur): ferme la fabrique de réponses perdues côté navigateur
Le magasin partait d'un dictionnaire de paramètres vide, et l'échec de la
requête qui devait le remplir restait sans conséquence visible. Les
options de la question du taux d'incapacité se calculent à partir de l'un
de ces paramètres : deux des trois valaient alors NaN — « De 50% à NaN% »
— et JSON écrit ce NaN `null`. La réponse était perdue à l'instant même
où elle était enregistrée.

La liste de repli rejoint `lib/` et sert de valeur initiale au magasin,
remplacée dès que la requête aboutit. Ce sont les mêmes valeurs que le
serveur sert déjà quand OpenFisca ne répond pas, et `computeParameter`
continue de signaler chaque recours. Aucune propriété ne peut plus
produire de NaN, pas seulement celle-ci.

Vérifié en navigateur, requête des paramètres en échec :
sans la valeur initiale, ["0.3","NaN","NaN"] puis `value: null` en base ;
avec, ["0.3","0.65","0.9"] puis `value: 0.9`.

`fieldsToAnswerAgain` porte deux rayons d'action très inégaux : la
migration ne touche que les simulations antérieures à la v18, tandis que
la garde des résultats renvoie dans le parcours toute simulation vivante
à qui la réponse manque. Un nom ajouté à cette liste sans migration
correspondante y renverrait aussitôt les simulations qui ne l'ont jamais
eue — l'avertissement est écrit auprès de la constante.
