+++ backend/lib/openfisca/parameters.ts
@@ -2,14 +2,10 @@ import * as Sentry from "@sentry/node"
 import openfisca from "./getter.js"
 
 import { OpenfiscaParameters } from "../../../lib/types/parameters.js"
+// La liste est partagée avec le navigateur, qui s'en sert de valeur initiale.
+import { parametersList } from "../../../lib/openfisca-parameters.js"
 
-export const parametersList: OpenfiscaParameters = {
-  "prestations_sociales.education.carte_des_metiers.age_maximal": 26,
-  "prestations_sociales.prestations_etat_de_sante.invalidite.aah.taux_capacite.taux_incapacite": 0.8,
-  "taxation_capital.epargne.livret_a.taux": 0.005,
-  "marche_travail.salaire_minimum.smic.smic_b_horaire": 10.57,
-  "marche_travail.salaire_minimum.smic.nb_heures_travail_mensuel": 151.67,
-}
+export { parametersList }
 
 // Délai avant qu'une tentative échouée puisse être relancée. Sans lui, chaque
 // appel de `getParameters` — un par droit calculé — repartirait sur le réseau
